feat: Enhance floorplan UI with improved labels, color legends, and automatic floorplan switching logic
