+++ frontend/.gitignore
@@ -28,6 +28,10 @@ yarn-error.log
 !.vscode/extensions.json
 .history/*
 
+#Extentions folder
+.claude/
+
+
 # Miscellaneous
 /.angular/cache
 .sass-cache/
