+++ frontend/src/app/features/floorplan/services/svg-color.service.ts
@@ -76,7 +76,7 @@ export class SvgColorService {
           (el as any).style.pointerEvents = 'auto';
         } else {
           el.setAttribute('fill', '#d1d5db'); // Light gray for unselected rooms (instead of 'none')
-          el.setAttribute('opacity', '0.35');
+          el.setAttribute('opacity', '0.10');
           (el as any).style.pointerEvents = 'auto';
         }
         (el as any).style.cursor = 'pointer';
